delete expiration_date from Plan database model
